FDCAN: Add separate file with structure definition
