+++ sw-mcu/canfd-to-usb-nucleo/Core/Src/utl_fdcan.c
@@ -6,14 +6,7 @@
  */
 
 #include "utl_fdcan.h"
-
-struct utl_fdcan_handle_t {
-  uint8_t payload[FDCAN_PAYLOAD];
-  FDCAN_HandleTypeDef handle;
-  FDCAN_TxHeaderTypeDef tx_header;
-  HAL_StatusTypeDef status_hal;
-  uint8_t id;
-};
+#include "utl_fdcan_definition.h"
 
 static struct utl_fdcan_handle_t fdcan_modules[FDCAN_COUNT];
